Reorganize interpreters/code_generators
