+++ m2cgen/ast/interpreters/__init__.py
@@ -1,4 +1,4 @@
-from .java import JavaInterpreter
+from .java.interpreter import JavaInterpreter
 
 __all__ = [
     JavaInterpreter
